cr, modify getSubtypes to be sync
