+++ packages/adapter-components/test/fetch/element/openAPI/invalid_openapi.yaml
