+++ packages/adapter-components/test/fetch/element/openAPI/petstore_openapi.v3.json
