+++ packages/adapter-components/test/fetch/element/openAPI/petstore_swagger.v2.json
